healthcare-chatbot/requirements.txt
streamlit
langchain
sentence-transformers
faiss-cpu
torch
transformers
python-dotenv
scikit-learn
tqdm
